- moved RunKafkaContainer to utils package - moved E2E test to separate test class
